jpg conversion ignored

+++ delete_contents.py
@@ -6,11 +6,13 @@ compress_pwd = "/Users/plangle-08/Player/compressed_images"   # directory for co
 for filename in os.listdir(jpg_pwd):
     file_path = os.path.join(jpg_pwd, filename)
     # print(file_path)
-    os.unlink(file_path)
+    if(file_path.endswith(".jpg" or ".png" or ".jpeg")):
+        os.unlink(file_path)
 print("DELETED CONVERTED_IMAGES")
 
 for filename in os.listdir(compress_pwd):
     file_path = os.path.join(compress_pwd, filename)
     #print(file_path)
-    os.unlink(file_path)
+    if (file_path.endswith(".jpg" or ".png" or ".jpeg")):
+        os.unlink(file_path)
 print("DELETED COMPRESSED_IMAGES")
